feat(viewing,favourite): implement full CRUD for viewing and favourite modules

+++ libs/database/src/entities/properties.entity.ts
@@ -50,7 +50,7 @@ export class PropertiesEntity extends BaseEntity {
   @Column({ type: 'int', nullable: true })
   year_built:number
   
-  @Column({ type: 'float', nullable: true })
+  @Column({ type: 'int', nullable: true })
   floors:number
 
   @Column({ type: 'int', nullable: true })
